CAL refactored
