+++ cal-rx_tx-demo/FakeECU/platformio.ini
@@ -13,7 +13,7 @@ platform = atmelavr
 board = uno
 framework = arduino
 lib_extra_dirs =
-    ../../cal/decoder
+    ../../cal/cal
 lib_deps =
-    cal-decoder
+    cal
     seeed-studio/CAN_BUS_Shield@^2.3.3
