+++ cal/OLD/encoder.cpp
