+++ cal/cal/examples/FakeECU/platformio.ini
@@ -12,11 +12,8 @@
 platform = atmelavr
 board = uno
 framework = arduino
-lib_extra_dirs = 
-	../../cal/cal
+lib_extra_dirs =
+    ../../
 lib_deps =
-	cal
-	seeed-studio/CAN_BUS_Shield@^2.3.3
-build_flags = -D
-check_tool = cppcheck
-check_flags = cppcheck: --addon=cert.py --supress=unusedFunction
+    cal
+    seeed-studio/CAN_BUS_Shield@^2.3.3
