+++ cal/cal/examples/Steering/platformio.ini
@@ -13,7 +13,7 @@ platform = atmelavr
 board = uno
 framework = arduino
 lib_extra_dirs = 
-	../../cal/cal
+	../../
 lib_deps =
 	cal
 	seeed-studio/CAN_BUS_Shield@^2.3.3
